Prepared test for floating numbers support

+++ src/model/Token.h
@@ -6,10 +6,9 @@
 #define EQUATION_SOLVER_TOKEN_H
 #include "TokenType.h"
 
-//TODO change into float and accept float numbers in the tokenizer
 struct Token {
     enum TokenType type;
-    int value;
+    float value;
 };
 
 #endif //EQUATION_SOLVER_TOKEN_H
